+++ stdpp/gmultiset.v
@@ -182,6 +182,24 @@ Section basic_lemmas.
     destruct X as [X]; simpl; rewrite elem_of_dom, <-not_eq_None_Some.
     destruct (X !! x); naive_solver lia.
   Qed.
+
+  Lemma gmultiset_Forall_dom P X : set_Forall P (dom X) ↔ set_Forall P X.
+  Proof. apply forall_proper; intros x. by rewrite gmultiset_elem_of_dom. Qed.
+  Lemma gmultiset_Exists_dom P X : set_Exists P (dom X) ↔ set_Exists P X.
+  Proof. apply exist_proper; intros x. by rewrite gmultiset_elem_of_dom. Qed.
+
+  Global Instance gmultiset_Forall_dec P X :
+    (∀ x, Decision (P x)) → Decision (set_Forall P X).
+  Proof.
+    refine (λ _, cast_if (decide (set_Forall P (dom X))));
+      by rewrite <-gmultiset_Forall_dom.
+  Defined.
+  Global Instance gmultiset_Exists_dec P X :
+    (∀ x, Decision (P x)) → Decision (set_Exists P X).
+  Proof.
+    refine (λ _, cast_if (decide (set_Exists P (dom X))));
+      by rewrite <-gmultiset_Exists_dom.
+  Defined.
 End basic_lemmas.
 
 (** * A solver for multisets *)